requirements.txt
hydra-core==1.3.2
numpy==2.4.1
omegaconf==2.3.0
pillow==12.0.0
pytorch-lightning==2.6.1
scikit-learn==1.8.0
sentence-transformers==5.2.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
timm==1.0.24
tqdm==4.67.3
transformers==5.1.0
wandb==0.25.0
